requirements.txt
Django==2.1.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyldap==3.0.0.post1
python-ldap==3.1.0
pytz==2018.7
